Add Hypothesis test that Distributions can compute in 32- and 64-bit floating point.
Also plumbing to make this work
- Fix a working precision problem in the SphericalUniform distribution
- Forbid nesting BatchReshape inside MixtureSameFamily in Hypothesis
  tests (b/161984806)
- Thread the `validate_args` argument through several Hypothesis strategies
  for Distributions and Bijectors

PiperOrigin-RevId: 333742615

+++ tensorflow_probability/python/distributions/spherical_uniform.py
@@ -158,7 +158,7 @@ class SphericalUniform(distribution.Distribution):
   def _log_prob(self, x):
     log_nsphere_surface_area = (
         np.log(2.) + (self.dimension / 2) * np.log(np.pi) -
-        tf.math.lgamma(tf.cast(self.dimension / 2., self.dtype)))
+        tf.math.lgamma(tf.cast(self.dimension / 2., x.dtype)))
     batch_shape = ps.broadcast_shape(
         ps.shape(x)[:-1], self.batch_shape)
     return tf.fill(batch_shape, -log_nsphere_surface_area)
